Add hand-crafted security token icon
Based on the material icon grid, in order to
match the dimensions of the material icons
'fingerprint', 'smartphone' and 'computer',
that we already use. The embedded shield
shapes are taken from the material
'security' icon.

The icon depicts a security icon with
a USB-C connector and the button
(e.g the touch button that people
are used to, from the yubikeys),
is represented as security shield.